requirements.txt
Django==2.0.5
djangorestframework==3.8.2
PyGithub==1.39
requests==2.18.4
slackclient==1.2.1
